refactor(buffer): extract inline test blocks into tests.rs and property_tests.rs
Moves the two #[cfg(test)] mod blocks (2,631 + 431 = 3,062 lines)
out of mod.rs into buffer/tests.rs and buffer/property_tests.rs.
mod.rs drops from 7,447 → 4,384 lines. Zero behaviour change —
tests remain in the same module path, use super::* still works.

127 buffer tests green.
